pyproject.toml
[project]
name = "jo-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain[openai]>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.19",
    "langchain-core>=0.3.74",
    "langchain-experimental>=0.3.4",
    "langchain-ollama>=0.3.6",
    "langchain-openai>=0.3.30",
    "langchain-postgres>=0.0.15",
    "langchain-text-splitters>=0.3.9",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.1.1",
    "ipykernel>=6.30.1",
    "streamlit>=1.48.1",
    "ctransformers>=0.2.27",
    "pypdf>=6.0.0",
    "pandas>=2.3.1",
    "streamlit-extras>=0.7.6",
    "faiss-cpu>=1.12.0",
]
